refactor: split `utils/constants` into new `constants` directory
