+++ src/app/constants/filesize.ts
@@ -1,0 +1,2 @@
+// 1 megabyte in bytes
+export const MB = 1048576
