+++ src/app/controllers/authentication.server.controller.js
@@ -25,7 +25,7 @@ const logger = require('../../config/logger').createLoggerWithLabel(
   'authentication',
 )
 const MailService = require('../services/mail.service').default
-const { EMAIL_HEADERS, EMAIL_TYPES } = require('../utils/constants')
+const { EMAIL_HEADERS, EMAIL_TYPES } = require('../constants/mail')
 
 const MAX_OTP_ATTEMPTS = 10
